pyproject.toml
[project]
name = "paligemmavlm-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.0",
    "numpy>=1.26.0,<2.1.0",
    "tensorflow-hub>=0.12.0",
    "tensorflow>=2.13.0,<2.19.0",
    "python-dotenv>=1.0.1",
    "pillow>=11.0.0",
    "matplotlib>=3.10.0",
    "keras-nlp>=0.18.1",
    "jax>=0.4.37",
]
